pyproject.toml
[project]
name = "pathseq"
version = "0.1.0"
description = "A pathlib-like library for working with file sequences."
authors = [
    {name = "Ashley Whetter", email = "ashley@awhetter.co.uk"},
]
readme = "README.rst"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Plugins",
    "Framework :: Sphinx :: Extension",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.9"
dependencies = [
    "lark",
    "python-statemachine",
    "typing_extensions",
]

[project.urls]
Source = "https://github.com/AWhetter/pathseq"
Documentation = "https://pathseq.readthedocs.org"

[dependency-groups]
dev = [
  {include-group = "test"},
  {include-group = "format"},
  {include-group = "lint"},
  {include-group = "typecheck"},
  {include-group = "doc"},
]
test = [
    "fileseq",
    "hypothesis",
    "pytest",
]
format = [
    "ruff",
]
lint = [
    "ruff",
]
typecheck = [
    "mypy",
]
doc = [
    "furo",
    "setuptools", # Missed dependency by kroki
    "sphinx",
    "sphinx-design",
    "sphinxcontrib-kroki",
]
release_notes = [
    "towncrier",
]

[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"

[tool.pylint]
disable = "R,unused-argument"

[tool.towncrier]
directory = "doc/changes"
filename = "CHANGELOG.rst"
package = "pathseq"
title_format = "pathseq v{version} ({project_date})"

[tool.pytest.ini_options]
markers = [
    "todo",
]
addopts = "--doctest-glob='*.rst'"
testpaths = [
    "tests",
    "doc/source",
]
